Code updates supporting client-only deployment
