+++ .gitignore
@@ -2,6 +2,8 @@
 node_modules
 /dist
 
+__pycache__
+
 # local env files
 .env.local
 .env.*.local
@@ -23,6 +25,9 @@ pnpm-debug.log*
 
 # Electron-builder output
 /dist_electron
+/dist
+
+/tools/mspfa/
 
 # install pseudofile
 install
