+++ Makefile
@@ -1,3 +1,5 @@
+-include .env
+
 CONFIG_JSON_PATH = "${APPDATA}/unofficial-homestuck-collection/config.json"
 
 default: test
