+++ package.json
@@ -52,9 +52,13 @@
     "mnemonist": "^0.39.5",
     "module": "^1.2.5",
     "node-loader": "^1.0.3",
+    "path-browserify": "^1.0.1",
+    "prettier": "^1.18.2",
+    "readline": "^1.3.0",
     "sass": "^1.62.1",
     "serve-handler": "^6.1.3",
     "sharp": "^0.32.1",
+    "string-replace-loader": "^3.1.0",
     "tar": "^6.1.11",
     "terser": "^4.8.0",
     "terser-webpack-plugin": "^3.0.3",
@@ -72,7 +76,9 @@
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
     "@babel/plugin-proposal-optional-chaining": "^7.21.0",
     "@babel/preset-env": "^7.21.4",
+    "@vue/cli-plugin-e2e-webdriverio": "^5.0.8",
     "@vue/cli-plugin-router": "~4.4.0",
+    "@vue/cli-plugin-unit-jest": "^5.0.8",
     "@vue/cli-service": "~4.4.0",
     "@vue/devtools": "^6.5.0",
     "@vue/eslint-config-standard": "^5.1.2",
@@ -87,11 +93,14 @@
     "eslint-plugin-promise": "^4.3.1",
     "eslint-plugin-standard": "^4.0.0",
     "eslint-plugin-vue": "^7.8.0",
+    "jest": "^29.5.0",
     "node-sass": "^7.0.1",
     "sass-loader": "^10.2.1",
     "vue-cli-plugin-electron-builder": "~2.1.1",
     "vue-template-compiler": "^2.7",
-    "webpack": "^4.18.0"
+    "webpack": "^4.18.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.15.1"
   },
   "resolutions": {
     "vue-cli-plugin-electron-builder/electron-builder": "^23.0.2"
